word-beautifier/requirements.txt
python-docx
tqdm
rich
